hf-space/requirements.txt
gradio==3.50.2
huggingface_hub<0.24
tencentcloud-sdk-python>=3.0.1200
